feat:UI 对接阶段 0 —— 串口状态镜像 → g_ui_sys
新 UI 原本和通信层零连接: g_ui_sys 是一组写死初值,
加热页还有模拟倒计时。本提交把数据面接上。

- 新增 lb_ui_sync_pull() (general_ui.c): 镜像 → g_ui_sys
  单向刷新, 挂在 func_process(), 页面只读
- 补 lb_get_display_tm() (lb_bridge.c): unix → 年月日时分,
  未同步时退回本机 RTC
- 补 fault_code 枚举 (lb_proto.h §4.1.6 全表);
  lowbat 改用 DP9 == LB_FAULT_LOW_BATTERY 判据, 不用 DP3
- time_min / keep_warm_min: u8 → u32 (保温 24H=1440 放不下),
  连带四处 snprintf 补 (unsigned) 强转
- 删掉 func_home_page.c 写死初值、func_heat_page.c 模拟
  倒计时与 remain_min 预置
- func_time_page.c 补 lb_time_set_synced(): 屏幕手改时间
  视为权威, 否则预约触发时间仍用 APP 旧基准
- 遗留事项新增 §2.1 映射表 / §2.2 待办(lid_open 无数据源、
  DP9 回值待实测、低电页是死胡同)

+++ projects/watch466/functions/new_ui/func_warm_page.c
@@ -190,10 +190,10 @@ static void func_warm_page_process(void)
             if (warm_min == 0) {
                 snprintf(buf, sizeof(buf), "0Min");
             } else if (warm_min % 60 == 0) {
-                snprintf(buf, sizeof(buf), "%uH", warm_min / 60);
+                snprintf(buf, sizeof(buf), "%uH", (unsigned)(warm_min / 60));
             } else {
                 snprintf(buf, sizeof(buf), "%uH%02uMin",
-                         warm_min / 60, warm_min % 60);
+                         (unsigned)(warm_min / 60), (unsigned)(warm_min % 60));
             }
             compo_textbox_set(inf->residue_time_text, buf);
         }
